requirements.txt
asgiref==3.8.1
Django==5.0.6
django-phone-field==1.8.1
django-phonenumber-field==8.0.0
gunicorn==22.0.0
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
setuptools==76.0.0
sqlparse==0.5.0
whitenoise==6.7.0
